Added date range search function
Added a function getDateRange() and a button which will toggle which data is seen. 

Button press will flip bool val stored in useDateRangeSwitch which determines what global data is used in the drawing chart functions. 

Changed listSiteData function to now take a bool, if true, the data passed into listSiteData will be checked for duplicates (times will be added together) otherwise duplicates are kept. This is for use in timeline so sites viewed on different days are not clumped together.